requirements.txt
addict==2.2.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
en-core-web-sm==2.2.5
entrypoints==0.3
faiss-cpu==1.6.1
gulpio==540.66
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==2.2.3
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
opencv-python==4.2.0.32
pandas==0.23.4
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycocotools==2.0.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
scipy==1.1.0
Send2Trash==1.5.0
sentencepiece==0.1.85
sh==1.12.14
six==1.14.0
spacy==2.2.3
srsly==1.0.1
style==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
torch==1.4.0
torchtext==0.5.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
update==0.0.1
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
zipp==2.2.0
